browser/pyproject.toml
[tool.poetry]
name = "browser"
version = "0.1.0"
description = ""
authors = ["Le Xuan Du <dulx96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
playwright = "^1.10.0"
aiorun = "^2020.12.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
